==> 5_predictions.ipynb <==
import pandas as pd
import numpy as np

from sklearn.linear_model import LinearRegression, BayesianRidge, ElasticNet
from sklearn.kernel_ridge import KernelRidge
from sklearn.ensemble import GradientBoostingRegressor
from sklearn.svm import SVR

from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error, mean_absolute_error


allModels = [LinearRegression(),
            KernelRidge(),
            ElasticNet(max_iter=2000),
            BayesianRidge(),
            GradientBoostingRegressor(),
            SVR()]

modelNames = ['Linear Regression',
              'Kernel Ridge',
              'Elastic Net',
              'Bayesian Ridge',
              'Gradient Boosting Regressor',
              'SVR']

# comment out to test different .csvs
df = pd.read_csv("csvs\\comb_money_cap.csv")
#df = pd.read_csv("csvs\\FinalDataSet_filtered.csv")

# -----------------------------------------------------------------
# was testing with income column
# income = df['REVENUE']-df['BUDGET']
# df['income']=income
#df = df.drop(columns=['MOVIE_ID', 'MOVIE_TITLE', 'RELEASE_YEAR', 'RELEASE_DATE', 'yt_releaseYear',
#                      'yt_releaseDate','REVENUE','BUDGET'])
# -----------------------------------------------------------------

df = df.drop(columns=['MOVIE_ID', 'MOVIE_TITLE', 'RELEASE_YEAR', 'RELEASE_DATE', 'yt_releaseYear',
                      'yt_releaseDate', 'BUDGET'])

print(df.columns)
print(df.shape)

# predict Tomatometer
X = df.drop('Tomatometer',axis= 1)
y = df['Tomatometer']

X_train, X_test, Y_train, Y_test = train_test_split(X,y,test_size=0.3, random_state=5)

for idx, model in enumerate(allModels):
    print(modelNames[idx])
    model.fit(X_train,Y_train)
    predictions_train = model.predict(X_train)
    predictions = model.predict(X_test)

    #print('predictions_train')
    #print('mean_squared_error : ', mean_squared_error(Y_train, predictions_train))
    #print('mean_absolute_error : ', mean_absolute_error(Y_train, predictions_train))

    print('mean_squared_error : ', mean_squared_error(Y_test, predictions))
    print('mean_absolute_error : ', mean_absolute_error(Y_test, predictions))

    testingRow = df.loc[[0]]
    print('truth: ' + str(testingRow['Tomatometer'][0]))

    testing = model.predict(testingRow.drop('Tomatometer',axis= 1))
    print('predict: ' + str(testing[0]))
    print()

# predict revenue (Didn'testingRowt work)
X = df.drop('REVENUE',axis= 1)
y = df['REVENUE']

X_train, X_test, Y_train, Y_test = train_test_split(X,y,test_size=0.3, random_state=5)

#print(X_train)

#from sklearn import preprocessing
#import numpy as np

#scaler = preprocessing.StandardScaler().fit(X_train)
#X_scaled = scaler.transform(df['REVENUE'])

for idx, model in enumerate(allModels):
    print(modelNames[idx])
    model.fit(X_train,Y_train)
    predictions_train = model.predict(X_train)
    predictions = model.predict(X_test)

    #print('predictions_train')
    #print('mean_squared_error : ', mean_squared_error(Y_train, predictions_train))
    #print('mean_absolute_error : ', mean_absolute_error(Y_train, predictions_train))

    #print('predictions_test')
    print('mean_squared_error : ', mean_squared_error(Y_test, predictions))
    print('mean_absolute_error : ', mean_absolute_error(Y_test, predictions))

    testingRow = df.loc[[0]]
    print('truth: ' + str(testingRow['REVENUE'][0]))

    testing = model.predict(testingRow.drop('REVENUE',axis= 1))
    print('predict: ' + str(testing[0]))
    print()
